Update nunjucksSetup to use filters/globals from helper file
